// src/app/core/models/clientesModel.ts
export class ClientesModel {
    id: number;
    razon_social: string;
    tip_ide: string;
    num_ide: string;
    tip_cont: string;
    nom_rep_leg: string;
    ape_rep_leg: string;
    tip_ide_rep_leg: string;
    num_ide_rep_leg: number;
    direcc: string;
    barrio: string;
    ciudad: number;
    depto: string;
    pais: string;
    tel_uno: string;
    tel_dos: string;
    fax: number;
    cod_postal: number;
    apto_aereo: number;
    mail: string;
    pag_web: string;
    asesor_comercial: string;
    forma_pago: string;
    term_pago: string;
    fact_a_nom: string;
    auto_retenedor: number;
    idunico: string;
    idemprendedor: number;
    state: number;
    cliente: string;
    constructor() {
        this.id = 0;
        this.razon_social = "";
        this.tip_ide = "";
        this.num_ide = "";
        this.tip_cont = "";
        this.nom_rep_leg = "";
        this.ape_rep_leg = "";
        this.tip_ide_rep_leg = "";
        this.num_ide_rep_leg = 0;
        this.direcc = "";
        this.barrio = "";
        this.ciudad = 0;
        this.depto = "";
        this.pais = "";
        this.tel_uno = "";
        this.tel_dos = "";
        this.fax = 0;
        this.cod_postal = 0;
        this.apto_aereo = 0;
        this.mail = "";
        this.pag_web = "";
        this.asesor_comercial = "";
        this.forma_pago = "";
        this.term_pago = "";
        this.fact_a_nom = "";
        this.auto_retenedor = 0;
        this.idunico = "";
        this.idemprendedor = 0;
        this.state = 0;
        this.cliente = "";
    }
}